+++ src/type-class/endo.ts
@@ -1,7 +1,7 @@
 import type { Monoid } from "./monoid";
 import { compose } from "../func";
 
-export type Endo<in out T> = (t: T) => T;
+export type Endo<T> = (t: T) => T;
 
 export const monoid = <T>(): Monoid<Endo<T>> => ({
     identity: (x: T) => x,
