+++ src/type-class/semi-group.ts
@@ -2,7 +2,7 @@
  * All instances of `SemiGroup` must satisfy following conditions:
  * - Associative: If `S` is a `SemiGroup`, for all `x`, `y` and `z`, `S.combine(S.combine(x, y), z)` equals to `S.combine(x, S.combine(y, z))`.
  */
-export interface SemiGroup<in out T> {
+export interface SemiGroup<T> {
     combine(l: T, r: T): T;
 }
