+++ src/writer.ts
@@ -1,13 +1,15 @@
+import * as Identity from "./identity";
+
 import type { GetHktA1, HktKeyA1 } from "./hkt";
 import type { Monad1, Monad2Monoid } from "./type-class/monad";
+
 import type { Functor2 } from "./type-class/functor";
-import { Identity } from "./lib";
 import type { Monoid } from "./type-class/monoid";
 
 declare const writerNominal: unique symbol;
 export type WriterHktKey = typeof writerNominal;
 
-export interface Writer<out W, out A> {
+export interface Writer<W, A> {
     (): [A, W];
 }
